Fixed random cone generating function
Before rotating to the provided axis, the random vector's reference
axis is +z.  It was passing the vector through unmodified if the
provided axis was +y (_2 = 1) when it should be +z (_3 = 1).  That
was probably from mis-porting something like vec[2] == 1.
Even after that, supplying +x and 0 deviation was returning -x.
Changed the literal vector from +z to -z when calculating the
rotation to the provided axis, presumably necessary due to some
handedness issue.  I can't remember which I ended up going with.
Left maybe?

Now there's some weird multiple reflections happening in the
shiny sphere.  Also something's up with the shadow of the matte
sphere.